Move internal to pkg to allow library re-use
