+++ cmd/oauth2.go
@@ -13,7 +13,7 @@ import (
 	"time"
 
 	"github.com/cli/browser"
-	"github.com/cloudentity/oauth2c/internal/oauth2"
+	"github.com/cloudentity/oauth2c/pkg/oauth2"
 	"github.com/imdario/mergo"
 	"github.com/spf13/cobra"
 )
