+++ cmd/oauth2_authorize_code.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/cloudentity/oauth2c/internal/oauth2"
 	"github.com/cli/browser"
+	"github.com/cloudentity/oauth2c/pkg/oauth2"
 )
 
 func (c *OAuth2Cmd) AuthorizationCodeGrantFlow(clientConfig oauth2.ClientConfig, serverConfig oauth2.ServerConfig, hc *http.Client) error {
