+++ cmd/oauth2_implicit.go
@@ -3,8 +3,8 @@ package cmd
 import (
 	"net/http"
 
-	"github.com/cloudentity/oauth2c/internal/oauth2"
 	"github.com/cli/browser"
+	"github.com/cloudentity/oauth2c/pkg/oauth2"
 )
 
 func (c *OAuth2Cmd) ImplicitGrantFlow(clientConfig oauth2.ClientConfig, serverConfig oauth2.ServerConfig, hc *http.Client) error {
