+++ pkg/oauth2/crypto_test.go
@@ -3,7 +3,7 @@ package oauth2_test
 import (
 	"testing"
 
-	"github.com/cloudentity/oauth2c/internal/oauth2"
+	"github.com/cloudentity/oauth2c/pkg/oauth2"
 	"github.com/stretchr/testify/require"
 )
